Cleanup workspace by creating window/class groups
