+++ MuzeiOSX/ImageSources/RedditSource.swift
@@ -1,57 +1,74 @@
 //
-//  WallpaperSource.swift
+//  RedditSource.swift
 //  MuzeiOSX
 //
-//  Created by Naman on 16/12/16.
+//  Created by Naman on 17/12/16.
 //  Copyright © 2016 naman14. All rights reserved.
 //
 
 import Foundation
-import Alamofire
 import SwiftyJSON
+import Alamofire
 import Kingfisher
 
-class FeaturedArtSource: WPSourceProtocol {
+class RedditSource: WallpaperSourceProtocol {
     
-    let API_URL: String = "https://muzeiapi.appspot.com/featured?cachebust=1"
+    let PREF_SUBREDDIT = "pref_subreddit"
     
-    func getWallpaper(callback: @escaping (Wallpaper) -> Void,  failure: @escaping () -> Void) {
+    var base_url = "https://www.reddit.com/r/{subreddit}/top.json?t=day&limit=10"
+    
+    func getWallpaper(callback: @escaping (Wallpaper) -> Void, failure: @escaping () -> Void) {
         
-        Alamofire.request(API_URL).responseJSON { response in
+        let prefs = UserDefaults.standard
+        
+        if let subreddit = prefs.string(forKey: PREF_SUBREDDIT) {
+            base_url = base_url.replacingOccurrences(of: "{subreddit}", with: subreddit)
+        } else {
+            base_url = base_url.replacingOccurrences(of: "{subreddit}", with: "EarthPorn")
+            
+        }
+        
+        Alamofire.request(base_url).responseJSON { (response) in
+            
             if let responseStr = response.result.value {
-                
                 let json = JSON(responseStr)
                 
-                let imageUri: String = json["imageUri"].stringValue
-                let title: String = json["title"].stringValue
+                let children = json["data"]["children"].arrayValue
                 
-                let byline = json["byline"].stringValue
-                let attribution = json["attribution"].stringValue
-                let detailsUri = json["detailsUri"].stringValue
+                let random = Int(arc4random_uniform(UInt32(5)))
 
-                let httpsUri = imageUri.replacingOccurrences(of: "http://", with: "https://", options: .literal, range: nil)
+                let data = children[random]["data"]
                 
+                let title = data["title"].stringValue
+                
+                let source = data["preview"]["images"][0]["source"]
+                
+                let imageUri = source["url"].stringValue
+                let byline = data["subreddit"].stringValue
+                let attribution = data["domain"].stringValue
+                let detailsUri = data["url"].stringValue
                 
                 if(ImageCache.default.imageCachedType(forKey: title).cached) {
+                    
                     ImageCache.default.retrieveImage(forKey: title, options: nil) {
                         image, cacheType in
                         if let image = image {
                             
                             let random = 2 + Int(arc4random_uniform(UInt32(999 - 2 + 1)))
                             
-                            let fileName: NSString = (imageUri as NSString).lastPathComponent as NSString
+                            let trimmedUrl =  NSURL(string: imageUri)?.absoluteStringByTrimmingQuery()
                             
-                            let fullURL = WPProcessor().getWallpaperFileUrl(
-                                fileName: fileName, processed: false, random: random)
+                            let fullURL = WallpaperProcessor().getWallpaperFileUrl(
+                                fileName: (trimmedUrl! as NSString).lastPathComponent as NSString, processed: false, random: random)
                             
-                            if WPProcessor().imageToFile(image: image, imageURL: fullURL!, ext:(imageUri as NSString).pathExtension) {
+                            if WallpaperProcessor().imageToFile(image: image, imageURL: fullURL!, ext:(trimmedUrl! as NSString).pathExtension) {
                                 
-                                let processedImage = WPProcessor().processImage(originalImage: image)
+                                let processedImage = WallpaperProcessor().processImage(originalImage: image)
                                 
-                                let processedURL = WPProcessor().getWallpaperFileUrl(
-                                    fileName: fileName, processed: true, random: random)
+                                let processedURL = WallpaperProcessor().getWallpaperFileUrl(
+                                    fileName: (trimmedUrl! as NSString).lastPathComponent as NSString, processed: true, random: random)
                                 
-                                if WPProcessor().imageToFile(image: processedImage, imageURL: processedURL!, ext:(imageUri as NSString).pathExtension) {
+                                if WallpaperProcessor().imageToFile(image: processedImage, imageURL: processedURL!, ext:(trimmedUrl! as NSString).pathExtension) {
                                     
                                     let wp: Wallpaper = Wallpaper(title: title,imageUrl: fullURL!, processedUrl: processedURL!)
                                     
@@ -73,15 +90,15 @@ class FeaturedArtSource: WPSourceProtocol {
                                 print("Error occurred")
                                 
                             }
-
+                            
                             
                         } else {
                             print("Not exist in cache.")
                         }
                     }
-                } else {
                     
-                    ImageDownloader.default.downloadImage(with: URL(string: httpsUri)!, options: [], progressBlock: nil) {
+                } else {
+                    ImageDownloader.default.downloadImage(with: URL(string: imageUri)!) {
                         (image, error, url, data) in
                         
                         if (error != nil) {
@@ -92,28 +109,28 @@ class FeaturedArtSource: WPSourceProtocol {
                         
                         let random = 2 + Int(arc4random_uniform(UInt32(999 - 2 + 1)))
                         
-                        let fileName: NSString = (imageUri as NSString).lastPathComponent as NSString
+                        let trimmedUrl =  NSURL(string: imageUri)?.absoluteStringByTrimmingQuery()
                         
-                        let fullURL = WPProcessor().getWallpaperFileUrl(
-                            fileName: fileName, processed: false, random: random)
                         
-                        if WPProcessor().imageToFile(image: image!, imageURL: fullURL!, ext:(imageUri as NSString).pathExtension) {
+                        let fullURL = WallpaperProcessor().getWallpaperFileUrl(
+                            fileName: (trimmedUrl! as NSString).lastPathComponent as NSString, processed: false, random: random)
+                        
+                        if WallpaperProcessor().imageToFile(image: image!, imageURL: fullURL!, ext:(trimmedUrl! as NSString).pathExtension) {
                             
-                            let processedImage = WPProcessor().processImage(originalImage: image!)
+                            let processedImage = WallpaperProcessor().processImage(originalImage: image!)
                             
-                            let processedURL = WPProcessor().getWallpaperFileUrl(
-                                fileName: fileName, processed: true, random: random)
+                            let processedURL = WallpaperProcessor().getWallpaperFileUrl(
+                                fileName: (trimmedUrl! as NSString).lastPathComponent as NSString, processed: true, random: random)
                             
-                            if WPProcessor().imageToFile(image: processedImage, imageURL: processedURL!, ext:(imageUri as NSString).pathExtension) {
+                            if WallpaperProcessor().imageToFile(image: processedImage, imageURL: processedURL!, ext:(trimmedUrl! as NSString).pathExtension) {
                                 
                                 let wp: Wallpaper = Wallpaper(title: title,imageUrl: fullURL!, processedUrl: processedURL!)
+                                ImageCache.default.store(image!, forKey: title)
                                 
                                 wp.byline = byline
                                 wp.attribution = attribution
                                 wp.detailsUri = detailsUri
-                                
-                                ImageCache.default.store(image!, forKey: title)
-                                
+
                                 callback(wp)
                                 
                             } else {
@@ -129,14 +146,17 @@ class FeaturedArtSource: WPSourceProtocol {
                             
                         }
                         
-                        
                     }
                 }
-                
             }
         }
         
         
     }
-    
 }
+
+
+
+
+
+
