+++ MuzeiOSX/WallpaperWindow/Wallpaper.swift
